#346: Move all mock classes into one place
