+++ src/app/components/layout/disclaimer-warning/disclaimer-warning.component.spec.ts
@@ -1,21 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NO_ERRORS_SCHEMA, Pipe, PipeTransform } from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { DisclaimerWarningComponent } from './disclaimer-warning.component';
 import { FeatureService } from '../../../services/feature.service';
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
-
-class MockFeatureService {
-  getFeatureToggleData(): any {
-    return {};
-  }
-}
+import { MockTranslatePipe, MockFeatureService } from '../../../utils/test-mocks';
 
 describe('DisclaimerWarningComponent', () => {
   let component: DisclaimerWarningComponent;
