+++ src/app/components/layout/double-button/double-button.component.spec.ts
@@ -1,14 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NO_ERRORS_SCHEMA, Pipe, PipeTransform } from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { DoubleButtonComponent } from './double-button.component';
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe } from '../../../utils/test-mocks';
 
 describe('DoubleButtonComponent', () => {
   let component: DoubleButtonComponent;
