+++ src/app/components/layout/header/header.component.spec.ts
@@ -1,46 +1,12 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NO_ERRORS_SCHEMA, Pipe, PipeTransform } from '@angular/core';
-import { Subject } from 'rxjs/Subject';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
-import { Observable } from 'rxjs/Observable';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { HeaderComponent } from './header.component';
 import { PriceService } from '../../../services/price.service';
 import { WalletService } from '../../../services/wallet.service';
 import { BlockchainService } from '../../../services/blockchain.service';
-import { TotalBalance } from '../../../app.datatypes';
 import { CoinService } from '../../../services/coin.service';
-import { BaseCoin } from '../../../coins/basecoin';
-
-class MockPriceService {
-  price: Subject<number> = new BehaviorSubject<number>(null);
-}
-
-class MockWalletService {
-  totalBalance: Subject<TotalBalance> = new BehaviorSubject<TotalBalance>(null);
-  hasPendingTransactions: Subject<boolean> = new ReplaySubject<boolean>();
-
-  sum() {
-  }
-}
-
-class MockBlockchainService {
-  get progress() {
-    return Observable.of();
-  }
-}
-
-class MockCoinService {
-  currentCoin = new BehaviorSubject<BaseCoin>(null);
-}
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe, MockCoinService, MockBlockchainService, MockPriceService, MockWalletService } from '../../../utils/test-mocks';
 
 describe('HeaderComponent', () => {
   let component: HeaderComponent;
