+++ src/app/components/layout/header/nav-bar/nav-bar.component.spec.ts
@@ -1,21 +1,11 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { Pipe, PipeTransform } from '@angular/core';
 import { MatIconModule, MatProgressSpinnerModule, MatTooltip } from '@angular/material';
 
 import { NavBarComponent } from './nav-bar.component';
 import { NavBarService } from '../../../../services/nav-bar.service';
 import { DoubleButtonComponent } from '../../double-button/double-button.component';
 import { ButtonComponent } from '../../button/button.component';
-
-class MockNavBarService {
-}
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe, MockNavBarService } from '../../../../utils/test-mocks';
 
 describe('NavBarComponent', () => {
   let component: NavBarComponent;
