+++ src/app/components/layout/loading-content/loading-content.component.spec.ts
@@ -1,14 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NO_ERRORS_SCHEMA, Pipe, PipeTransform } from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { LoadingContentComponent } from './loading-content.component';
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe } from '../../../utils/test-mocks';
 
 describe('LoadingContentComponent', () => {
   let component: LoadingContentComponent;
