+++ src/app/components/pages/buy/add-deposit-address/add-deposit-address.component.spec.ts
@@ -7,22 +7,7 @@ import { Observable } from 'rxjs/Observable';
 import { AddDepositAddressComponent } from './add-deposit-address.component';
 import { WalletService } from '../../../../services/wallet.service';
 import { PurchaseService } from '../../../../services/purchase.service';
-
-class MockWalletService {
-  get addresses(): Observable<any[]> {
-    return Observable.of([]);
-  }
-}
-
-class MockPurchaseService {
-}
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe, MockPurchaseService, MockWalletService } from '../../../../utils/test-mocks';
 
 describe('AddDepositAddressComponent', () => {
   let component: AddDepositAddressComponent;
