+++ src/app/components/pages/buy/buy.component.spec.ts
@@ -1,37 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatCardModule, MatDividerModule, MatIconModule, MatListModule, MatDialogModule } from '@angular/material';
-import { Pipe, PipeTransform } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
 
 import { BuyComponent } from './buy.component';
 import { PurchaseService } from '../../../services/purchase.service';
-
-class MockPurchaseService {
-  all(): Observable<any[]> {
-    return Observable.of([]);
-  }
-}
-
-@Pipe({name: 'tellerStatus'})
-class MockTellerStatusPipe implements PipeTransform {
-  transform() {
-    return 'transformed value';
-  }
-}
-
-@Pipe({ name: 'dateTime' })
-class MockDateTimePipe implements PipeTransform {
-  transform() {
-    return 'transformed value';
-  }
-}
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe, MockPurchaseService, MockTellerStatusPipe, MockDateTimePipe } from '../../../utils/test-mocks';
 
 describe('BuyComponent', () => {
   let component: BuyComponent;
