+++ src/app/components/pages/history/transaction-detail/transaction-detail.component.spec.ts
@@ -1,22 +1,10 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
-import { Subject } from 'rxjs/Subject';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
-import { NO_ERRORS_SCHEMA, Pipe, PipeTransform } from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { TransactionDetailComponent } from './transaction-detail.component';
 import { PriceService } from '../../../../services/price.service';
-
-class MockPriceService {
-  price: Subject<number> = new BehaviorSubject<number>(null);
-}
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe, MockPriceService } from '../../../../utils/test-mocks';
 
 describe('TransactionDetailComponent', () => {
   let component: TransactionDetailComponent;
