+++ src/app/components/pages/onboarding/onboarding-create-wallet/onboarding-disclaimer/onboarding-disclaimer.component.spec.ts
@@ -1,15 +1,9 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NO_ERRORS_SCHEMA, Pipe, PipeTransform } from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { MatDialogRef } from '@angular/material';
 
 import { OnboardingDisclaimerComponent } from './onboarding-disclaimer.component';
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe } from '../../../../../utils/test-mocks';
 
 describe('OnboardingDisclaimerComponent', () => {
   let component: OnboardingDisclaimerComponent;
