+++ src/app/components/pages/send-skycoin/send-form/send-form.component.spec.ts
@@ -1,46 +1,12 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatSnackBarModule, MatDialog } from '@angular/material';
-import { NO_ERRORS_SCHEMA, PipeTransform, Pipe } from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
-import { Observable } from 'rxjs/Observable';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 
 import { SendFormComponent } from './send-form.component';
 import { WalletService } from '../../../../services/wallet.service';
-import { Wallet } from '../../../../app.datatypes';
-import { BaseCoin } from '../../../../coins/basecoin';
 import { CoinService } from '../../../../services/coin.service';
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
-
-class MockWalletService {
-  get all(): Observable<Wallet[]> {
-    return Observable.of([]);
-  }
-}
-
-class MockCoinService {
-  currentCoin = new BehaviorSubject<BaseCoin>({
-    id: 1,
-    nodeUrl: 'nodeUrl',
-    nodeVersion: 'v1',
-    coinName: 'test coin',
-    coinSymbol: 'test',
-    hoursName: 'test',
-    cmcTickerId: 1,
-    coinExplorer: 'testUrl'
-  });
-}
-
-class MockMatSnackBar {
-  dismiss() {
-  }
-}
+import { MockTranslatePipe, MockWalletService, MockMatSnackBar, MockCoinService } from '../../../../utils/test-mocks';
 
 describe('SendFormComponent', () => {
   let component: SendFormComponent;
