+++ src/app/components/pages/settings/outputs/outputs.component.spec.ts
@@ -1,31 +1,13 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NO_ERRORS_SCHEMA, Pipe, PipeTransform } from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { ActivatedRoute } from '@angular/router';
 import { MatDialogModule } from '@angular/material';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 
 import { OutputsComponent } from './outputs.component';
 import { WalletService } from '../../../../services/wallet.service';
-import { BaseCoin } from '../../../../coins/basecoin';
 import { CoinService } from '../../../../services/coin.service';
-
-class MockWalletService {
-  outputsWithWallets() {
-    return Observable.of([]);
-  }
-}
-
-class MockCoinService {
-  currentCoin = new BehaviorSubject<BaseCoin>(null);
-}
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe, MockWalletService, MockCoinService } from '../../../../utils/test-mocks';
 
 describe('OutputsComponent', () => {
   let component: OutputsComponent;
