+++ src/app/components/pages/wallets/change-name/change-name.component.spec.ts
@@ -1,20 +1,11 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { NO_ERRORS_SCHEMA, Pipe, PipeTransform } from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
 import { FormBuilder } from '@angular/forms';
 
 import { ChangeNameComponent } from './change-name.component';
 import { WalletService } from '../../../../services/wallet.service';
-
-class MockWalletService {
-}
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
+import { MockTranslatePipe, MockWalletService } from '../../../../utils/test-mocks';
 
 describe('ChangeNameComponent', () => {
   let component: ChangeNameComponent;
