+++ src/app/components/pages/wallets/wallet-detail/wallet-detail.component.spec.ts
@@ -1,31 +1,14 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatDialog, MatSnackBar, MatMenuModule } from '@angular/material';
 import { RouterTestingModule } from '@angular/router/testing';
-import { NO_ERRORS_SCHEMA, Pipe, PipeTransform } from '@angular/core';
+import { NO_ERRORS_SCHEMA } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
-import { Observable } from 'rxjs/Observable';
 
 import { WalletDetailComponent } from './wallet-detail.component';
 import { WalletService } from '../../../../services/wallet.service';
-
-@Pipe({name: 'translate'})
-class MockTranslatePipe implements PipeTransform {
-  transform() {
-    return 'translated value';
-  }
-}
-
-class MockWalletService {
-}
-
-class MockTranslateService {
-  get(key: string | Array<string>, interpolateParams?: Object): Observable<string | any> {
-    return Observable.of({});
-  }
-}
+import { MockTranslatePipe, MockWalletService, MockTranslateService } from '../../../../utils/test-mocks';
 
 describe('WalletDetailComponent', () => {
-
   let component: WalletDetailComponent;
   let fixture: ComponentFixture<WalletDetailComponent>;
