+++ src/app/directives/clipboard.directive.spec.ts
@@ -1,18 +1,9 @@
 import { TestBed, inject } from '@angular/core/testing';
-import { Component } from '@angular/core';
 import { By } from '@angular/platform-browser';
 
 import { ClipboardDirective } from './clipboard.directive';
 import { ClipboardService } from '../services/clipboard.service';
-
-@Component({
-  template: `<span clipboard="test data"></span>`
-})
-class TestComponent {
-}
-
-class MockClipboardService {
-}
+import { MockClipboardService, TestComponent } from '../utils/test-mocks';
 
 describe('Directive: Clipboard', () => {
   beforeEach(() => {
