+++ src/app/services/api.service.spec.ts
@@ -1,16 +1,10 @@
 import { TestBed, inject } from '@angular/core/testing';
 import { MockBackend } from '@angular/http/testing';
 import { HttpModule, XHRBackend } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
-
-import { ApiService } from './api.service';
 import { TranslateService } from '@ngx-translate/core';
 
-class MockTranslateService {
-  get(key: string | Array<string>, interpolateParams?: Object): Observable<string | any> {
-    return Observable.of({});
-  }
-}
+import { ApiService } from './api.service';
+import { MockTranslateService } from '../utils/test-mocks';
 
 describe('ApiService', () => {
   let service: ApiService;
